# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Urva Gandhi Portfolio Developer API & MCP Server",
    "description": "Public REST APIs, Model Context Protocol (MCP) server, and developer endpoints providing verified coding statistics, developer profile metadata, and interactive contact handling for Urva Gandhi.",
    "version": "1.0.0",
    "contact": {
      "name": "Urva Gandhi",
      "email": "23bce078@nirmauni.ac.in",
      "url": "https://urvagandhi.tech"
    },
    "x-versioning-policy": "URL path versioning (/v1/) and X-API-Version header parameter. Backward compatibility guaranteed for version 1.",
    "x-deprecation-policy": "Deprecation notices served via Sunset HTTP header 6 months prior to end-of-life."
  },
  "servers": [
    {
      "url": "https://urvagandhi.tech",
      "description": "Primary Production Domain"
    },
    {
      "url": "https://urvagandhi-portfolio.vercel.app",
      "description": "Lifetime Vercel Mirror Domain"
    },
    {
      "url": "https://urvagandhi.tech/v1",
      "description": "Active Domain Versioned Base URL (/v1)"
    }
  ],
  "paths": {
...
    "/api/codeforces": {
      "get": {
        "operationId": "getCodeforcesStats",
        "summary": "Get Codeforces Statistics",
        "description": "Fetches Codeforces contest rating, max rating, rank, and submission metrics.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "Codeforces handle (defaults to Urva_Gandhi)",
            "schema": {
              "type": "string",
              "default": "Urva_Gandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Codeforces response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeforcesProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/codeforces": {
      "get": {
        "operationId": "getCodeforcesStats",
        "summary": "Get Codeforces Statistics",
        "description": "Fetches Codeforces contest rating, max rating, rank, and submission metrics.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "Codeforces handle (defaults to Urva_Gandhi)",
            "schema": {
              "type": "string",
              "default": "Urva_Gandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Codeforces response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeforcesProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/codeforces": {
      "get": {
        "operationId": "getCodeforcesStats",
        "summary": "Get Codeforces Statistics",
        "description": "Fetches Codeforces contest rating, max rating, rank, and submission metrics.",
        "security": [],
        "parameters": [
          {
            "name": "X-API-Version",
            "in": "header",
            "required": false,
            "description": "API version to target. Omit to use the latest stable version (1).",
            "schema": {
              "type": "string",
              "enum": ["1"],
              "default": "1"
            }
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "description": "Codeforces handle (defaults to Urva_Gandhi)",
            "schema": {
              "type": "string",
              "default": "Urva_Gandhi"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Codeforces response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeforcesProfile"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          },
          "authorizationCode": {
            "authorizationUrl": "https://urvagandhi.tech/.well-known/oauth/authorize",
            "tokenUrl": "https://urvagandhi.tech/.well-known/oauth/token",
            "scopes": {
              "profile:read": "Read public developer profile and coding statistics.",
              "profile:write": "Submit a contact inquiry on behalf of an authenticated client."
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["type", "title", "status", "code", "detail"],
        "properties": {
          "type": {
            "type": "string",
            "example": "https://urvagandhi.tech/docs/errors/invalid-parameters"
          },
          "title": {
            "type": "string",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "MISSING_REQUIRED_FIELDS"
          },
          "detail": {
            "type": "string",
            "example": "Name, email, and message fields are required."
          },
          "instance": {
            "type": "string",
            "example": "/api/contact"
          },
          "resolution_hint": {
            "type": "string",
            "example": "Ensure request body contains valid name, email, and message."
          }
        }
      },
...
      "CodeforcesProfile": {
        "type": "object",
        "properties": {
          "handle": {
            "type": "string",
            "example": "Urva_Gandhi"
          },
          "rating": {
            "type": "integer",
            "example": 1350
          },
          "maxRating": {
            "type": "integer",
            "example": 1420
          },
          "rank": {
            "type": "string",
            "example": "pupil"
          }
        }
      },
...
    }
  }
}